+++ scripts/database/otrs-initial_insert.postgresql.sql
@@ -748,6 +748,12 @@ INSERT INTO ticket_history_type (name, valid_id, create_by, create_time, change_
     VALUES
     ('EmailResend', 1, 1, current_timestamp, 1, current_timestamp);
 -- ----------------------------------------------------------
+--  insert into table ticket_history_type
+-- ----------------------------------------------------------
+INSERT INTO ticket_history_type (name, valid_id, create_by, create_time, change_by, change_time)
+    VALUES
+    ('Bulk', 1, 1, current_timestamp, 1, current_timestamp);
+-- ----------------------------------------------------------
 --  insert into table article_sender_type
 -- ----------------------------------------------------------
 INSERT INTO article_sender_type (name, valid_id, create_by, create_time, change_by, change_time)